+++ topobank/manager/urls.py
@@ -5,10 +5,9 @@ from rest_framework.routers import DefaultRouter
 from . import views
 
 router = DefaultRouter()
-router.register(r'api/tag', views.TagViewSet, basename='tag-api')
-router.register(r'api/topography', views.TopographyViewSet, basename='topography-api')
-router.register(r'api/surface', views.SurfaceViewSet, basename='surface-api')
-router.register(r'api/property', views.PropertyViewSet, basename='property-api')
+router.register(r"api/tag", views.TagViewSet, basename="tag-api")
+router.register(r"api/topography", views.TopographyViewSet, basename="topography-api")
+router.register(r"api/surface", views.SurfaceViewSet, basename="surface-api")
 
 urlpatterns = router.urls
 
@@ -22,53 +21,46 @@ urlpatterns += [
     # Data routes
     #
     re_path(
-        r'api/surface/(?P<surface_id>\d+)/download/$',
+        r"api/surface/(?P<surface_id>\d+)/download/$",
         view=views.download_surface,
-        name='surface-download'
+        name="surface-download",
     ),
     #
     # API routes
     #
     path(
-        'api/topography/<pk>/force-inspect/',
+        "api/topography/<pk>/force-inspect/",
         view=login_required(views.force_inspect),
-        name='force-inspect'
+        name="force-inspect",
     ),
     path(
-        'api/surface/<pk>/set-permissions/',
-        view=login_required(views.set_permissions),
-        name='set-permissions'
+        "api/surface/<pk>/set-permissions/",
+        view=login_required(views.set_surface_permissions),
+        name="set-surface-permissions",
+    ),
+    re_path(
+        r"api/set-tag-permissions/<(?P<name>[^.]+)>",
+        view=login_required(views.set_tag_permissions),
+        name="set-tag-permissions",
     ),
     path(
-        'api/tag-numerical-properties/<pk>/',
+        "api/tag-numerical-properties/<pk>/",
         view=login_required(views.tag_numerical_properties),
-        name='numerical-properties'
+        name="numerical-properties",
     ),
     path(
-        'api/tag-categorical-properties/<pk>/',
+        "api/tag-categorical-properties/<pk>/",
         view=login_required(views.tag_categorical_properties),
-        name='categorical-properties'
+        name="categorical-properties",
     ),
     path(
-        'api/import-surface/',
+        "api/import-surface/",
         view=login_required(views.import_surface),
-        name='import-surface'
+        name="import-surface",
     ),
-    path(
-        'api/versions/',
-        view=views.versions,
-        name='versions'
-    ),
-    path(
-        'api/statistics/',
-        view=views.statistics,
-        name='statistics'
-    ),
+    path("api/versions/", view=views.versions, name="versions"),
+    path("api/statistics/", view=views.statistics, name="statistics"),
     # GET
     # * Return memory usage of inspection tasks
-    path(
-        'api/memory-usage/',
-        view=views.memory_usage,
-        name='memory-usage'
-    )
+    path("api/memory-usage/", view=views.memory_usage, name="memory-usage"),
 ]